+++ Mediscreen-Note/src/main/java/com/bintou/mediscreen/note/repository/NoteRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface NoteRepository extends MongoRepository<Note, String> {
 
-    List<Note> findByPatientId(final Integer patientId);
+    List<Note> findByPatientId(final Long patientId);
 }
